requirements.txt
streamlit
tensorflow-cpu
pillow
numpy
opencv-python-headless
